requirements.txt
fastapi
uvicorn
python-docx
PyPDF2
sentence-transformers
faiss-cpu
requests
python-dotenv